+++ src/main/java/com/github/neuralnetworks/architecture/activation/ActivationFunction.java
@@ -6,5 +6,5 @@ package com.github.neuralnetworks.architecture.activation;
  *
  */
 public interface ActivationFunction {
-	public double value(double input);
+	public float value(float input);
 }
